[Feat] 마이페이지 notifier과 provider 분리 (#123)
