+++ lib/view/my_page/component/my_vege_stack_box.dart
@@ -1,7 +1,7 @@
 import 'package:farmus/common/farmus_theme_color.dart';
 import 'package:farmus/view/my_page/component/my_history_box.dart';
 import 'package:farmus/view/my_page/component/my_page_veges.dart';
-import 'package:farmus/view_model/my_page/notifier/vege_count_notifier.dart';
+import 'package:farmus/view_model/my_page/vege_count_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
