+++ lib/view/my_page/my_farmclub/component/my_farmclub_stack_box.dart
@@ -1,7 +1,7 @@
 import 'package:farmus/common/farmus_theme_color.dart';
 import 'package:farmus/view/my_page/my_farmclub/component/farmclub_widget.dart';
 import 'package:farmus/view/my_page/my_farmclub/component/my_farmclub_box.dart';
-import 'package:farmus/view_model/my_page/notifier/vege_count_notifier.dart';
+import 'package:farmus/view_model/my_page/vege_count_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
