+++ lib/view/my_page/vege_history/component/vege_info.dart
@@ -1,0 +1,15 @@
+class VegeInfo {
+  String veggieName;
+  String veggieType;
+  String periodStart;
+  String periodEnd;
+  String image;
+
+  VegeInfo({
+    required this.veggieName,
+    required this.veggieType,
+    required this.periodStart,
+    required this.periodEnd,
+    required this.image,
+  });
+}
