+++ lib/view_model/my_page/notifier/vege_count_notifier.dart
@@ -1,9 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-final vegeCountProvider = StateNotifierProvider<VegeCountNotifier, int>((ref) {
-  return VegeCountNotifier();
-});
-
 class VegeCountNotifier extends StateNotifier<int> {
   VegeCountNotifier() : super(0);
