+++ lib/view_model/my_page/notifier/vege_profile_notifier.dart
@@ -1,20 +1,5 @@
+import 'package:farmus/view/my_page/vege_history/component/vege_info.dart';
 import 'package:flutter/material.dart';
-
-class VegeInfo {
-  String veggieName;
-  String veggieType;
-  String periodStart;
-  String periodEnd;
-  String image;
-
-  VegeInfo({
-    required this.veggieName,
-    required this.veggieType,
-    required this.periodStart,
-    required this.periodEnd,
-    required this.image,
-  });
-}
 
 class VegeInfoNotifier extends ChangeNotifier {
   VegeInfo _vegeInfo = VegeInfo(
